initial

+++ dashboard-admin.php
@@ -1,35 +1,37 @@
 <?php
-    include "config/connection.php";
+include "config/connection.php";
 
-    session_start();
-    include 'config/connection.php';
+session_start();
+include 'config/connection.php';
 
 
-    if(!isset($_SESSION["username"])) {
-        header('Location: login.php?status=restrictedaccess');
-        exit;
+if (!isset($_SESSION["username"])) {
+    header('Location: login.php?status=restrictedaccess');
+    exit;
+}
+
+
+
+function rupiah($angka)
+{
+    $hasil_rupiah = "Rp. " . number_format($angka, 0, '.', '.');
+    return $hasil_rupiah;
+}
+function query($query)
+{
+    global $conn;
+    $result = mysqli_query($conn, $query);
+    $rows = [];
+    while ($row = mysqli_fetch_assoc($result)) {
+        $rows[] = $row;
     }
+    return $rows;
+}
 
-    
+// WHERE status_donasi = 'Diterima'
 
-    function rupiah($angka){
-        $hasil_rupiah = "Rp. ".number_format($angka,0,'.','.');
-        return $hasil_rupiah;
-    }
-   function query($query){
-       global $conn;
-        $result = mysqli_query($conn,$query); 
-        $rows = [];
-        while($row = mysqli_fetch_assoc($result)){
-            $rows[] = $row;
-        }
-        return $rows;
-   }
-
-    // WHERE status_donasi = 'Diterima'
-
-    // var_dump($programDonasi);die;
-   $programDonasi = query("SELECT *, SUM(t_donasi.nominal_donasi) AS dana_terkumpul_total, 
+// var_dump($programDonasi);die;
+$programDonasi = query("SELECT *, SUM(t_donasi.nominal_donasi) AS dana_terkumpul_total, 
                     COUNT(id_user) 
                     AS jumlah_donatur 
                     FROM t_donasi 
@@ -50,7 +52,7 @@
 
 
 //    $programDonasi = query("SELECT * FROM t_program_donasi");
-    
+
 ?>
 
 
@@ -74,7 +76,7 @@
     <link rel="stylesheet" type="text/css" href="css/dashboard-yst.css">
     <!-- Google Fonts -->
     <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link href="https://fonts.googleapis.com/css2?family=Nunito+Sans:wght@400;600;700&family=Roboto:wght@500&display=swap" rel="stylesheet"> 
+    <link href="https://fonts.googleapis.com/css2?family=Nunito+Sans:wght@400;600;700&family=Roboto:wght@500&display=swap" rel="stylesheet">
 </head>
 
 <body class="hold-transition sidebar-mini layout-fixed">
@@ -88,16 +90,15 @@
                 </li>
             </ul>
             <!-- Right navbar links -->
-            <ul class="navbar-nav ml-auto user-wrapper"> 
+            <ul class="navbar-nav ml-auto user-wrapper">
                 <img src="img/user-default.jpg" width="30px" height="30px" alt="">
-                <li class="nav-item dropdown user-dropdown">  
-                    <a class="nav-link dropdown-toggle pr-4" href="#" id="navbarDropdownMenuLink" 
-                    role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                    <?php echo("{$_SESSION['username']}");?>
+                <li class="nav-item dropdown user-dropdown">
+                    <a class="nav-link dropdown-toggle pr-4" href="#" id="navbarDropdownMenuLink" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                        <?php echo ("{$_SESSION['username']}"); ?>
                     </a>
-                    <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">      
+                    <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
                         <a class="dropdown-item" href="logout.php">Logout</a>
-                    </div>                   
+                    </div>
                 </li>
             </ul>
         </nav>
@@ -108,13 +109,13 @@
             <!-- Brand Logo -->
 
             <a href="dashboard-admin.php" class="brand-link">
-                <img src="img/logo-only.svg"  class="brand-image mt-1">
+                <img src="img/logo-only.svg" class="brand-image mt-1">
                 <span class="brand-text font-weight-bold mt-2"><i>Dashboard Admin</i></span>
             </a>
 
             <!-- Sidebar -->
             <div class="sidebar">
-           <!-- Sidebar Menu -->
+                <!-- Sidebar Menu -->
                 <nav class="mt-4">
                     <ul class="nav nav-pills nav-sidebar flex-column" data-widget="treeview" role="menu" data-accordion="false">
                         <!-- Add icons to the links using the .nav-icon class
@@ -154,10 +155,18 @@
                             </a>
                         </li>
                         <li class="nav-item nav-item-sidebar">
+                            <a href="kelola-berita.php" class="nav-link side-icon">
+                                <i class="nav-icon fas fa-newspaper"></i>
+                                <p>
+                                    Kelola Berita
+                                </p>
+                            </a>
+                        </li>
+                        <li class="nav-item nav-item-sidebar">
                             <a href="laporan-program-donasi.php" class="nav-link side-icon">
                                 <i class="nav-icon fas fa-calendar-check"></i>
                                 <p>
-                                    Lp.  Program Donasi
+                                    Lp. Program Donasi
                                 </p>
                             </a>
                         </li>
@@ -173,7 +182,7 @@
                             <a href="laporan-program-relawan.php" class="nav-link side-icon">
                                 <i class="nav-icon fas fa-calendar-check"></i>
                                 <p>
-                                    Lp.  Program Relawan
+                                    Lp. Program Relawan
                                 </p>
                             </a>
                         </li>
@@ -181,11 +190,11 @@
                             <a href="laporan-relawan.php" class="nav-link side-icon ">
                                 <i class="nav-icon fas fa-address-book"></i>
                                 <p>
-                                    Lp.  Relawan
+                                    Lp. Relawan
                                 </p>
                             </a>
                         </li>
-                        
+
                     </ul>
                 </nav>
                 <!-- /.sidebar-menu -->
@@ -196,31 +205,31 @@
         <!-- Content Wrapper. Contains page content -->
         <div class="content-wrapper">
             <main>
-            <div class="request-data">
+                <div class="request-data">
                     <div class="projects">
-                        <div class="page-title-link ml-4 mb-4">     
+                        <div class="page-title-link ml-4 mb-4">
                             <a href="dashboard-admin.php">
-                                <i class="nav-icon fas fa-home mr-1"></i>Dashboard admin</a> > 
+                                <i class="nav-icon fas fa-home mr-1"></i>Dashboard admin</a> >
                             <a href="dashboard-admin.php">
                                 <i class="nav-icon fas fa-cog mr-1"></i>Program donasi</a>
                         </div>
 
                         <div class="card card-request-data">
                             <div class="card-header-req">
-                                 <div class="row ml-1 ">
+                                <div class="row ml-1 ">
                                     <div class="col ">
-                                      <div class="dropdown show ">
-                                        <a class="btn btn-info  filter-btn dropdown-toggle" href="#" role="button" id="dropdownMenuLink" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                                          Filter
-                                        </a>
-                                        <div class="dropdown-menu green-drop" aria-labelledby="dropdownMenuLink">
-                                          <a class="dropdown-item" href="status-berjalan.php">Berjalan</a>
+                                        <div class="dropdown show ">
+                                            <a class="btn btn-info  filter-btn dropdown-toggle" href="#" role="button" id="dropdownMenuLink" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                                                Filter
+                                            </a>
+                                            <div class="dropdown-menu green-drop" aria-labelledby="dropdownMenuLink">
+                                                <a class="dropdown-item" href="status-berjalan.php">Berjalan</a>
+                                            </div>
                                         </div>
                                     </div>
-                                    </div>
-                              </div>
-                              <button class="mr-5" onclick="location.href='input-program-donasi.php'">Input Program Donasi <span class="fas fa-plus-square"></span></button>
-    
+                                </div>
+                                <button class="mr-5" onclick="location.href='input-program-donasi.php'">Input Program Donasi <span class="fas fa-plus-square"></span></button>
+
                             </div>
                             <div class="card-body card-body-req">
                                 <div class="table-responsive">
@@ -232,46 +241,45 @@
                                                 <td>Dana Terkumpul</td>
                                                 <td>Target Dana</td>
                                                 <td class="text-center">Jumlah <br> Donatur</td>
-                                                <td class="text-center">Status<br>  Program</td>
-                                                <td class="justify-content-center" >Aksi</td>
+                                                <td class="text-center">Status<br> Program</td>
+                                                <td class="justify-content-center">Aksi</td>
                                             </tr>
                                         </thead>
                                         <tbody>
-                                        <?php foreach($programDonasi as $row):?>
-                                            <tr>
-                                                <td class="text-center"><?= $row["id_program_donasi"]; ?></td>
-                                                <td class="table-snipet1"><?= $row["nama_program_donasi"]; ?></td>
-                                                <td><?= rupiah($row['dana_terkumpul_total']) == 0 ? '0' : rupiah($row['dana_terkumpul_total']); ?></td>
-                                                <td><?= rupiah($row["target_dana"]); ?></td>
-                                                <td class="text-center"><?= $row["jumlah_donatur"]; ?></td>
-                                                <td class="text-center">
-                                                    <?= $row["status_program_donasi"]; ?>
-                                                </td>
-                                                <td class="justify-content-between">
-                                                    <button type="button" class="btn btn-edit">
-                                                        <a href="edit-program-donasi.php?id_program_donasi=<?= $row["id_program_donasi"]; ?>" class="fas fa-edit"></a>
-                                                    </button>
+                                            <?php foreach ($programDonasi as $row) : ?>
+                                                <tr>
+                                                    <td class="text-center"><?= $row["id_program_donasi"]; ?></td>
+                                                    <td class="table-snipet1"><?= $row["nama_program_donasi"]; ?></td>
+                                                    <td><?= rupiah($row['dana_terkumpul_total']) == 0 ? '0' : rupiah($row['dana_terkumpul_total']); ?></td>
+                                                    <td><?= rupiah($row["target_dana"]); ?></td>
+                                                    <td class="text-center"><?= $row["jumlah_donatur"]; ?></td>
+                                                    <td class="text-center">
+                                                        <?= $row["status_program_donasi"]; ?>
+                                                    </td>
+                                                    <td class="justify-content-between">
+                                                        <button type="button" class="btn btn-edit">
+                                                            <a href="edit-program-donasi.php?id_program_donasi=<?= $row["id_program_donasi"]; ?>" class="fas fa-edit"></a>
+                                                        </button>
 
-                                                    <!-- <php if($row['status_program_donasi'] == 'Siap disalurkan' || $row['status_program_donasi'] == 'Selesai'){?>
+                                                        <!-- <php if($row['status_program_donasi'] == 'Siap disalurkan' || $row['status_program_donasi'] == 'Selesai'){?>
                                                     <button type="button" class="btn btn-edit">
                                                         <a href="edit-program-donasi.php?id_program_donasi=<= $row["id_program_donasi"]; ?>" class="fa fa-upload"></a>
                                                     </button>
                                                     <php } ?> -->
-                                                    
-                                                    <button type="button" class="btn btn-delete ml-1">
-                                                        <a href="hapus.php?type=pdonasi&id_program_donasi=<?=$row["id_program_donasi"];?>" 
-                                                        class="far fa-trash-alt" onclick="return confirm('Anda yakin ingin menghapus program ini ?');"></a>
-                                                    </button>
-                                                  </td>
-                                            </tr>
-                                        <?php endforeach;?>
+
+                                                        <button type="button" class="btn btn-delete ml-1">
+                                                            <a href="hapus.php?type=pdonasi&id_program_donasi=<?= $row["id_program_donasi"]; ?>" class="far fa-trash-alt" onclick="return confirm('Anda yakin ingin menghapus program ini ?');"></a>
+                                                        </button>
+                                                    </td>
+                                                </tr>
+                                            <?php endforeach; ?>
                                         </tbody>
                                     </table>
                                 </div>
                             </div>
                         </div>
                     </div>
-                    
+
                 </div>
             </main>
         </div>
@@ -279,7 +287,7 @@
         <!-- /.content -->
     </div>
     <!-- /.content-wrapper -->
-    
+
     <footer class="main-footer">
         <center><strong> &copy; YST 2021.</strong> Yayasan Sekar Telkom </center>
     </footer>
