+++ kelola-donasi.php
@@ -1,35 +1,37 @@
 <?php
 
-    session_start();
-    include 'config/connection.php';
+session_start();
+include 'config/connection.php';
 
 
-    if(!isset($_SESSION["username"])) {
-        header('Location: login.php?status=restrictedaccess');
-        exit;
-    }
+if (!isset($_SESSION["username"])) {
+    header('Location: login.php?status=restrictedaccess');
+    exit;
+}
 
-    function rupiah($angka){
-        $hasil_rupiah = "Rp. ".number_format($angka,0,'.','.');
-        return $hasil_rupiah;
-    }
+function rupiah($angka)
+{
+    $hasil_rupiah = "Rp. " . number_format($angka, 0, '.', '.');
+    return $hasil_rupiah;
+}
 
-    function query($query){
-       global $conn;
-        $result = mysqli_query($conn, "SELECT * FROM t_donasi
+function query($query)
+{
+    global $conn;
+    $result = mysqli_query($conn, "SELECT * FROM t_donasi
                       LEFT JOIN t_program_donasi ON t_donasi.id_program_donasi = t_program_donasi.id_program_donasi
                       ORDER BY id_donasi DESC
-                        "); 
-        $rows = [];
-        while($row = mysqli_fetch_assoc($result)){
-            $rows[] = $row;
-        }
-        return $rows;
+                        ");
+    $rows = [];
+    while ($row = mysqli_fetch_assoc($result)) {
+        $rows[] = $row;
     }
+    return $rows;
+}
 
 
-   $result = query("SELECT * FROM t_donasi");
-   
+$result = query("SELECT * FROM t_donasi");
+
 //    var_dump($donasiSaya);die;
 
 ?>
@@ -54,7 +56,7 @@
     <link rel="stylesheet" type="text/css" href="css/dashboard-yst.css">
     <!-- Google Fonts -->
     <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link href="https://fonts.googleapis.com/css2?family=Nunito+Sans:wght@400;600;700&family=Roboto:wght@500&display=swap" rel="stylesheet"> 
+    <link href="https://fonts.googleapis.com/css2?family=Nunito+Sans:wght@400;600;700&family=Roboto:wght@500&display=swap" rel="stylesheet">
 </head>
 
 <body class="hold-transition sidebar-mini layout-fixed">
@@ -68,16 +70,15 @@
                 </li>
             </ul>
             <!-- Right navbar links -->
-            <ul class="navbar-nav ml-auto user-wrapper"> 
+            <ul class="navbar-nav ml-auto user-wrapper">
                 <img src="img/user-default.jpg" width="30px" height="30px" alt="">
-                <li class="nav-item dropdown user-dropdown">  
-                    <a class="nav-link dropdown-toggle pr-4" href="#" id="navbarDropdownMenuLink" 
-                    role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                    <?php echo("{$_SESSION['username']}");?>
+                <li class="nav-item dropdown user-dropdown">
+                    <a class="nav-link dropdown-toggle pr-4" href="#" id="navbarDropdownMenuLink" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                        <?php echo ("{$_SESSION['username']}"); ?>
                     </a>
-                    <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">      
+                    <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
                         <a class="dropdown-item" href="logout.php">Logout</a>
-                    </div>                   
+                    </div>
                 </li>
             </ul>
         </nav>
@@ -88,13 +89,13 @@
             <!-- Brand Logo -->
 
             <a href="dashboard-admin.php" class="brand-link">
-                <img src="img/logo-only.svg"  class="brand-image mt-1">
+                <img src="img/logo-only.svg" class="brand-image mt-1">
                 <span class="brand-text font-weight-bold mt-2"><i>Dashboard Admin</i></span>
             </a>
 
             <!-- Sidebar -->
             <div class="sidebar">
-           <!-- Sidebar Menu -->
+                <!-- Sidebar Menu -->
                 <nav class="mt-4">
                     <ul class="nav nav-pills nav-sidebar flex-column" data-widget="treeview" role="menu" data-accordion="false">
                         <!-- Add icons to the links using the .nav-icon class
@@ -134,10 +135,18 @@
                             </a>
                         </li>
                         <li class="nav-item nav-item-sidebar">
+                            <a href="kelola-berita.php" class="nav-link side-icon">
+                                <i class="nav-icon fas fa-newspaper"></i>
+                                <p>
+                                    Kelola Berita
+                                </p>
+                            </a>
+                        </li>
+                        <li class="nav-item nav-item-sidebar">
                             <a href="laporan-program-donasi.php" class="nav-link side-icon">
                                 <i class="nav-icon fas fa-calendar-check"></i>
                                 <p>
-                                    Lp.  Program Donasi
+                                    Lp. Program Donasi
                                 </p>
                             </a>
                         </li>
@@ -153,7 +162,7 @@
                             <a href="laporan-program-relawan.php" class="nav-link side-icon">
                                 <i class="nav-icon fas fa-calendar-check"></i>
                                 <p>
-                                    Lp.  Program Relawan
+                                    Lp. Program Relawan
                                 </p>
                             </a>
                         </li>
@@ -161,11 +170,11 @@
                             <a href="laporan-relawan.php" class="nav-link side-icon ">
                                 <i class="nav-icon fas fa-address-book"></i>
                                 <p>
-                                    Lp.  Relawan
+                                    Lp. Relawan
                                 </p>
                             </a>
                         </li>
-                        
+
                     </ul>
                 </nav>
                 <!-- /.sidebar-menu -->
@@ -175,12 +184,12 @@
 
         <!-- Content Wrapper. Contains page content -->
         <div class="content-wrapper">
-        <main>
-            <div class="request-data">
+            <main>
+                <div class="request-data">
                     <div class="projects">
-                        <div class="page-title-link ml-4 mb-4">     
+                        <div class="page-title-link ml-4 mb-4">
                             <a href="dashboard-admin.php">
-                                <i class="nav-icon fas fa-home mr-1"></i>Dashboard admin</a> > 
+                                <i class="nav-icon fas fa-home mr-1"></i>Dashboard admin</a> >
                             <a href="kelola-donasi.php">
                                 <i class="nav-icon fas fa-donate mr-1"></i>Kelola donasi</a>
                         </div>
@@ -189,62 +198,60 @@
                             <div class="card-header-req">
                                 <div class="row ml-1 ">
                                     <div class="col ">
-                                      <div class="dropdown show ">
-                                        <a class="btn btn-info  filter-btn dropdown-toggle" href="#" role="button" id="dropdownMenuLink" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                                          Filter
-                                        </a>
-                                        <div class="dropdown-menu green-drop" aria-labelledby="dropdownMenuLink">
-                                          <a class="dropdown-item" href="#">Tampilkan Semua</a>
-                                         
-                                      </div>
+                                        <div class="dropdown show ">
+                                            <a class="btn btn-info  filter-btn dropdown-toggle" href="#" role="button" id="dropdownMenuLink" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                                                Filter
+                                            </a>
+                                            <div class="dropdown-menu green-drop" aria-labelledby="dropdownMenuLink">
+                                                <a class="dropdown-item" href="#">Tampilkan Semua</a>
+
+                                            </div>
+                                        </div>
                                     </div>
-                                    </div>
-                              </div>
-                        
+                                </div>
+
                             </div>
                             <div class="card-body card-body-req">
                                 <div class="table-responsive">
                                     <table width="100%" class="text-center">
                                         <thead>
-                                            <tr >
+                                            <tr>
                                                 <td>Kode<br>Donasi</td>
                                                 <td>Nama Donatur</td>
                                                 <td>Tgl Donasi</td>
                                                 <td>Nominal</td>
-                                                <td >Kode Program <br>Pilihan</td>
+                                                <td>Kode Program <br>Pilihan</td>
                                                 <td>Status Donasi</td>
-                                                <td class="justify-content-center" >Aksi</td>
+                                                <td class="justify-content-center">Aksi</td>
                                             </tr>
                                         </thead>
-                                        <tbody> 
-                                        <?php foreach($result as $row):?>                            
-                                            <tr>
-                                                <td><?= $row["id_donasi"]; ?></td>
-                                                <td><?= $row["nama_donatur"]; ?></td>
-                                                <td><?= $row["tgl_donasi"]; ?></td>
-                                                <td ><?= rupiah($row["belum_dibayar"]); ?></td>
-                                                <!-- <td class="table-snipet1"> -->
-                                                <td class="text-center"><?= $row["id_program_donasi"]; ?></td>
-                                                <td><?= $row["status_donasi"]; ?></td>
-                                                <td class="justify-content-center">
-                                                    <button type="button" class="btn btn-edit">
-                                                        <a href="edit-donasi.php?id_donasi=<?= $row["id_donasi"]; ?>" 
-                                                        class="fas fa-edit"></a>
-                                                    </button>
-                                                    <button type="button" class="btn btn-delete ml-1">
-                                                        <a href="hapus.php?type=donasi&id_donasi=<?=$row["id_donasi"];?>" 
-                                                        class="far fa-trash-alt" onclick="return confirm('Anda yakin ingin menghapus donasi ini ?');"></a>
-                                                    </button>
-                                                  </td>
-                                            </tr> 
-                                        <?php endforeach;?>   
+                                        <tbody>
+                                            <?php foreach ($result as $row) : ?>
+                                                <tr>
+                                                    <td><?= $row["id_donasi"]; ?></td>
+                                                    <td><?= $row["nama_donatur"]; ?></td>
+                                                    <td><?= $row["tgl_donasi"]; ?></td>
+                                                    <td><?= rupiah($row["belum_dibayar"]); ?></td>
+                                                    <!-- <td class="table-snipet1"> -->
+                                                    <td class="text-center"><?= $row["id_program_donasi"]; ?></td>
+                                                    <td><?= $row["status_donasi"]; ?></td>
+                                                    <td class="justify-content-center">
+                                                        <button type="button" class="btn btn-edit">
+                                                            <a href="edit-donasi.php?id_donasi=<?= $row["id_donasi"]; ?>" class="fas fa-edit"></a>
+                                                        </button>
+                                                        <button type="button" class="btn btn-delete ml-1">
+                                                            <a href="hapus.php?type=donasi&id_donasi=<?= $row["id_donasi"]; ?>" class="far fa-trash-alt" onclick="return confirm('Anda yakin ingin menghapus donasi ini ?');"></a>
+                                                        </button>
+                                                    </td>
+                                                </tr>
+                                            <?php endforeach; ?>
                                         </tbody>
                                     </table>
                                 </div>
                             </div>
                         </div>
                     </div>
-                    
+
                 </div>
             </main>
         </div>
@@ -252,7 +259,7 @@
         <!-- /.content -->
     </div>
     <!-- /.content-wrapper -->
-    
+
     <footer class="main-footer">
         <center><strong> &copy; YST 2021.</strong> Yayasan Sekar Telkom </center>
     </footer>
